Connect form
We were getting the following error:

"Form submission canceled because the form is not connected"

The answer to this was to move the buttons
out of the form tags and into the card actions.

Functionally the form is very similar to the login form,
and so I changed it to be more visually consistent.