Fix Categories selection
This changes the selection so that the user is not able to create new categories, this will be
added later as an explicit action that creates the category in the db first, so that it can then
be added to the monitor. This change also stores the selected category list in the correct
type.